% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expometa.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It should be the same with the PID generated by InitBioLink.}
}
\value{
Exit status
}
\description{
End the module analysis
}
\examples{
res = InitMeta()
   res1 = LoadMeta(PID = res$PID,
   UseExample = "example#1",
   DataPath = NULL)
   res2 = MetaEffect(PID=res$PID,
   OutPath = "default",
   CID = "default",
   DID = "default")
   FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
