% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expometa.R
\name{InitMeta}
\alias{InitMeta}
\title{Initialize ExpoMeta Module}
\usage{
InitMeta()
}
\value{
An R6 class object.
}
\description{
Initialize ExpoMeta module, the first step to start ExpoMeta Module.
}
\details{
ExpoMeta module mainly provides users preliminary information retrieval and screening for meta-analysis. Run InitMeta to get a unique PID for following steps.
}
\examples{
res = InitMeta()
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
