% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expometa.R
\name{LoadMeta}
\alias{LoadMeta}
\title{Load Data for ExpoMeta Module}
\usage{
LoadMeta(PID = NULL, Example = "default", DataPath = NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitMeta.}

\item{UseExample}{chr. A character indicates whether uses example data for analyses, available option include "example#1" for using example data1 and "default" for using data uploaded.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_meta.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Upload local data file for ExpoMeta Module.
}
\details{
After initializing ExpoMeta module, the second step is to upload local data file for ExpoMeta Module.
Set param "UseExample = 'example#1'" to use example data1.
LoadMeta can only run successfully after successfully running InitMeta.
Please attention, PID must be got from the return result of InitMeta().
}
\examples{
res = InitMeta()
   res1 = LoadMeta(PID = res$PID,
   UseExample = "example#1",
   DataPath = NULL)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
