% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expometa.R
\name{VizRefer}
\alias{VizRefer}
\title{Visualize the Articles' Main Information}
\usage{
VizRefer(PID = NULL, OutPath = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitMeta.}

\item{OutPath}{chr. Output file directory, e.g. "D:/output". If "default", the current working directory will be set. It should be noted that the slash symbol is "/", not "\".}
}
\value{
A list object containing plots of article information visualization.
}
\description{
Visualize the articles' main information after MetaRefer function.
}
\details{
VizRefer provides the visual function of articles' main information, showing the year and region distribution directly.
Please attention, PID must be got from the return result of InitMeta().
VizRefer can only run successfully after successfully running InitMeta, LoadMeta and MetaRefer functions.
}
\examples{
res = InitMeta()
   res1 = LoadMeta(PID = res$PID,
   UseExample = "example#1",
   DataPath = NULL)
   res2 = MetaRefer(PID = res$PID,
   OutPath = "default",
   Mode = "search",
   VarX = "default",
   VarY = "default",
   VarM = "default",
   YearFrom = "default",
   earEnd = "default",
   PMID = "default")
   res3 = VizRefer(PID=res$PID,
   OutPath = "default")
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
